// src/app/core/interface/dadosForm.interface.ts
export interface AberturaForm {
    co_cep: number;
    co_entidade_registro: number;
    co_municipio: string;
    co_numero: string;
    co_uf: string;
    date_nascimento: string;
    ds_bairro: string;
    ds_complemento: string;
    ds_logradouro: string;
    ds_nome_fantasia: string;
    ds_responsavel: string;
    nu_cpf: string;
    co_natureza_juridica: number;
    id: number;

}